Example of sixel hardcopy from VT340 in "Compressed" mode (DECGEPM reset).
Note that the image is offset to the right by 50 pixels. This is
because this particular file was created before mediacopy.sh handled
removing the default printing margin.

Also note that the image has 800x240 pixels, but should display as
800x480 due to a pixel aspect ratio of 2:1. Currently libsixel,
ImageMagick, and XTerm do not handle this correctly.

Since the VT340 screen is 800x480 pixels, a "compressed" print loses
resolution in the Y-axis.