Improve output by using Sixel Level 2 Graphics
Default output (Sixel Level 1 graphics) uses a 2:1 pixel aspect ratio,
which worked for the DEC printers of the time. If we use Level 2
graphics, we get the proper 1:1 aspect ratio (square pixels). Even
better, a lot of the weird bugs in how the sixels are being sent go
away. Most importantly, now we can rely on ST (Esc \) to be received
at the end of the Media Copy.