Add sample output from running the Media Copy to host command.
These .six files are minimally edited from the sixel graphics data
returned by the VT340 when executing Print Screen. The main change is
that the first two characters (Esc \) have been removed as it causes
libsixel and ImageMagick to get confused and think the image has only
one pixel.

The filename reflect the options used during printing:

Level 1 vs Level 2 sixel graphics
Compressed or Expanded printing
Normal or Rotated printing
Fullscreen or Cropped

Note that these tests show that these options are not orthogonal.
Compressed or Expanded printing only makes sense for Level 1.

The option to make the background transparent (which was always on in
all of the current samples) only works with Level 2 prints.

Level 1 hard copy uses a 2:1 pixel aspect ratio (to match the printers
of the time) and that's why it uses the wacky expanded/compressed
kludges.

Level 2 hard copy uses a 1:1 pixel aspect ratio, which makes more
sense, but needs to be changed by hand in the VT340 Set-Up menu.