Glitchy screenshot from color_selection.sh
In addition to the glitch at the top due to ImageMagick/libsixel,
there is a glitch about 2/3rds of the way down in this image which is
perhaps caused by an error in how the VT340 sends sixel data during a
hardcopy.

The Media Copy command, which took about a 90 seconds for the other
screenshots, took about 400 seconds because it paused for a long time
where the glitch occurs. Shortly before the pause, the VT340 sends a
character with the eighth bit high, which should never happen in a
sixel sequence. It appears to be a reproducible bug, but it is
peculiar that it only occurs on the image created by color_selection.sh.

Could it be one of the escape sequences being sent is invalid and
putting the terminal in a confused state? Or is it something about the
data shown on the screen? Note that the other .sh tests, which ran
after this test, a soft reset, and resetpalette, did not exhibit the
glitch.