Raw sixel from color_selection.sh for debugging 8-bit glitch
This is raw output from the VT340 Media Copy command as it was
returned when running color_selection.sh in 80- and 132-column mode.
The glitch did not occur in 132-column mode. (Well, not exactly raw as
I removed the ST preamble (Esc \) which was confusing ImageMagick.)

color_selection.six-part is the partial data received at the point
that the VT340 paused sending data for about five minutes. Note that
there is already a character with the eighth bit set, so whatever went
wrong happened before the pause.

color_selection.six is the full data received from the VT340 after
about 7 minutes.

color_selection-132.six is the data received by running
color_selection.sh in 132-column mode. There was no glitch.