Make sure the background fill is triggered at the start of the image
The background fill only begins on the first line that outputs a sixel
data value. So if the image starts with one or more DECGNL commands,
they will need to be prefixed with at least one data value (potentially
zero) in order to trigger a fill from the very top of the image.