ST kludge no longer needed
When transferring images in the VT340's default "compressed" mode, we
never received the final line of data which meant the String
Terminator:(Esc \) was never received. That would leave a terminal
stuck in sixel mode, if catted to the screen, so we had been appending
ST to the file. Now that we use expanded mode, that is no longer an issue.