A script to test whether DECSDM is implemented the same as the VT340.
This is important because the documentation for the VT340 has this reversed
as do any terminals based on that flawed documentation.

On a genuine VT340, when DECSDM is reset, sixels scroll the screen,
and when DECSDM is set, sixels do not scroll.

We can detect this by checking the text cursor positioning. When
sixels scroll, the terminal is supposed to put the text cursor on
the last line of the sixel image. When sixels do not scroll, the
text cursor is supposed to be unaffected by any images.

We send an image 60 pixels high, which hopefully is taller than
anyone's screen font height, and see if the cursor position has
changed. We do that with DECSDM both SET and RESET and then report
whether it is the same or not as a true VT340 would do.

____

What this is not:

This script does *not* report other differences that could be found
with this test.

Other test scripts should be written to check these examples:

Some sixel parsers send a carriage return, sending the cursor position
to the first column, after sixel data. The original VT340 does not do
that.

Some sixel parsers send a new line after sixel data so that the next
line of text will not overwrite the last line of graphics. The VT340
does not do that (although that is a misfeature I consider a bug).

Some sixel parsers will not move the cursor at all if there are no
pixels being drawn. (For example, only graphics carriage returns). The
VT340 does move the cursor.

Some sixel parsers will move the cursor down based on the geometry
hint given in the sixel header, regardless of the actual sixel data.
The VT340 leaves the cursor at the end of the sixel data and ignores
the header.