Thank you Paul Flo Williams for the scan of Text Programming, 2nd Edition!
Paul, who runs vt100.net, happened to have this book in his cupboard
and kindly scanned it when I asked.

I have run tesseract OCR on this version and edited the PDF page
numbering to match the actual page numbers.

The OCR could definitely be improved as it detected a lot of
"diacritical marks", but this seems good enough for now.