Reset color palette by switching to map 0 then map 1 (glitchy)
Causes text glitches occasionally.

A rectangle of text that had previously been on the screen reappears,
sometimes even triggering scrolling. Happens more when running the
test-decstglt.sh script and (possibly?) when using emacs. It is not at
all clear what causes certain text to be copied or why switching
colormaps would make it reappear. However, the same text is pasted
every time if the glitch is triggered repeatedly.