Modified tek.c to output Tek4010/4014 demo to screen
This uses ANSI escape sequences to print text in colors on the screen,
which is a feature of the MS Kermit terminal emulator. While the
colors do not work, the code has been modified so it does not
completely fail on a VT340 terminal. It just prints the escape
sequences as numbers on the screen.

Also, added a -p option to take a screenshot using the Tek print
command. This relies on the MediaCopy capabilities of the VT340
terminal which can redirect printing to the host instead of the
printer.