Improve the image_movement test
This is an attempt to improve the test output in the ICH/DCH portion of
the test. Ideally it should look reasonable on both a real VT340, which
erases the affected portion of the line, as well as on modern terminals,
which may support the movement of image data along with the text.