Improve the line_attributes test
This is an attempt to improve the output of the test so that it looks
reasonable on both a real VT340, as well as on modern terminals, that
aren't necessarily limited in the same way. I've also added a section
showing how sixel could be used to trick the VT340 into displaying
both single-width and double-height text on the same line.