Screenshot of output from tek.c using the -P option.
Printing a media copy to the host is buggy on the VT340 when using the
Tektronix print command. Half of the output gets splatted to the
screen instead of to the host. Instead, this image was captured by
hooking up another serial port on the host to the VT340's printer port
and reading the normal "print to printer" media copy command.