Fix reading bug by using Esc as delimiter
Previously I had been looking for the backslash (\) in the String
Terminator to find the end of the Media Copy. I had somehow believed
that \ was not a valid character in sixel data, but of course it
actually is. Esc is not valid data, so now I'm splitting on that.

Splitting lines on Esc also makes it easy to return just the DCS
string that contains the sixel data, and not the preliminary stuff
that the VT340 sends: String Terminator, Carriage Return, and a
sequence to set the image resolution:

	 Esc \ CR Esc [ 2 SP I

While those are completely valid in a sixel image file, they confuse
simplistic programs, like ImageMagick.