Lotka-Volterra predator-prey model
Equations for the 'ode' program to pipe into 'graph' to create ReGIS
graphics output for the VT340.

This ode file draws a phase portrait for the so-called Lotka-Volterra
predator-prey model. Suppose that in a lake there are two species of
fish: A (the prey) who live by eating a plentiful supply of plants,
and B (the predator) who eat A. Let x(t) be the population of A and
y(t) the population of B at time t. A crude model for the interaction
of A and B is given by the equations

     x' = x(a-by)
     y' = y(cx-d)

where a, b, c, d are positive constants.

To graph:      ode < lotka-volterra | graph -T regis -C -x 0 5 -y 0 5