Lorenz strange attractor equations
The Lorenz model, a system of three coupled ordinary differential
equations with parameter r, one of the most famous _strange
attractors_ in mathematics.

To graph on a VT340:

   ode < lorenz | graph -T regis -C -x -10 10 -y -10 10

It takes a couple minutes to finish on a genuine VT340, but is more
enjoyable to watch in progress than one might expect.