Add targets to run ODE examples for ReGIS
ODE is an Ordinary Differential Equation solver that comes with the
plotutils packages. It can be piped into the graph program to produce
ReGIS output. However, the command lines to do so are a bit messy, so
this Makefile exists so that they can be run easily:

     make lo  Lorenz Attractor
     make at  Swinging Atwood's Machine
     make lv  Lotka-Volterra Predator/Prey Model

One can also save the ReGIS output to files by using make regisoutput.
Then the *.regis files can be simply cat'ed to the screen to view.