Handtweaked sixel version of the |d|i|g|i|t|a|l| logo
As specified by the logo designer's notes, the lettering is actually
negative space, holes where the background shows through.

This uses sixel color #9 without defining it, so it should usually be
the system's default (#565699), which is a nice dark blue. By not
redefining the colormap, this logo should work no matter what the
user's favorite color scheme is.

By the way, because color register #0 (black) is always the screen
background color on the VT340, we cannot use it as the color to draw
with. The only way to draw in black on a different color screen
background would be to redefine one of the other colors, #1 to #15, to
be black.