Clarify that DCS (Esc P) is always 7-bit from MediaCopy
The receive_media_copy() function takes a shortcut by using getdelim
to chop up the input on the Escape character. This is guaranteed to be
okay by the documentation for the VT340 which states that even in
8-bit mode, Esc P will still be used for the sixel output. (I have
tested this and it is true.)

If ever this turns out to be a problem, we can replace getdelim with a
simple statemachine, such as I wrote in signalhandler.c which reads a
byte at a time. That code handles 7-bit CSI (Esc [) as well as 8-bit
CSI (0x9B).

By the way, regardless of mode chosen in Set-up -> General -- 7-bit or
8-bit -- the VT340 will always display sixel images, whether they use
7-bit (Esc P) or 8-bit DCS (0x90) and 7-bit (Esc \) or 8-bit ST (0x9C).