Sample sixel file from saitoha/libsixel/images
Note that this file does not end with a newline or graphic newline and
is only six pixels high. After sending it to the console, the text
cursor will be on the same line as the image and will overwrite it,
character by character, if any text is printed. See map8.png for an example.