Initial commit of color inversion - cursor blink test
Demonstrates color inversion used by the VT340 when blinking the
cursor. It uses the fg color's index exclusive or'd with 7 (0111) to
get the inverse color's index.

That gives us these inverse pairs:

        SATURATED     MUTED COLORS
	0-------7   8-------------15
	 1-----6      9---------14
	  2---5	       10-----13
	   3-4		 11-12

Note that the saturated colors from 0 to 7 are inverted with each
other, instead of mixing with the muted colors from 8 to 15.

Looking at the color instead of the index, when using the default
VT340 palette the inverse color can be described succinctly in HLS
colorspace:

    If Saturation is 0 (grays),
	Lightness Percent is offset by 50 percentile points.

    If Saturation > 0 (colors),
	Hue Angle is rotated by 180 degrees.