Add blink+bold (index 8) text colour
Previous script had a stub saying to add in the correct ID once the
"VT340 blink FCO" arrives. The correct ID is 8. Whatever color is set
in index 8 is used as the foreground color for text that is bold when
blinking.

Normal, non-bold, blinking text does not have a separate color ID
since it simply swaps the foreground and background colors at 1 Hz.