Screenshot from rgb.sh
Not visible is that the blinking text is, once a second, swapping fg
and bg (changing from sky blue text on a midnight blue background to
midnight blue on sky blue).

Also not visible: the bold blinking text is toggling betwen white on
midnight blue and red on sky blue.