Latency of Primary DA on VT340 is shocking high
This script shows that it takes over a fifth of a second for the VT340
to respond to ESC [ c (Primary DA). That matters because hackerb9 had
been using it as a way to time the VT340. In the future, Device Status
Request 5 (Hardware check) will be more accurate as it takes only
0.03s.