Use regex instead of Bash's broken read for parsing
Yes, an ANSI parser would make more sense, but I don't know of a handy
command line one. (Recommendations welcome.)

Bash 5.2's read is broken because I was reading the file directly into
variables with integer type. I had thought it wouldn't be a problem
because anything that's not an integer gets turned into 0, right?
Nope!

Turns out this script crashes with a syntax error and no way to handle
it, when the input has a slash in it:

   declare -i i
   read i

As a workaround, I had been parsing the data twice, once in a subshell
to check for errors before doing it for real. It was very silly:

    # Sanity check: Read data into integers first in a subshell as certain
    # input could cause bash to die with a syntax error. E.g., '#!/bin/bash'.
    if ! ( IFS=$';' read Pfn Pcn Pe Pcmw Pw Pt Pcmh PcssDscsSxbp1 Sxbp2 <<<"$data" >/dev/null 2>&1 ); then
	echo "Error parsing $filename" >&2
	return 1
    fi

Parsing with regular expressions is of course still not correct, but
it is a lot better than a kludge like that.