Testing locale transliteration. Does not work, yet. Maybe ever.
My current understanding is that one needs to write a special gconv
module and compile it for each machine to get interactive
transliteration working. (E.g., running a program that uses APL
symbols in Unicode and shows them on a VT340.)

I thought maybe making a custom transliteration table in the locale
definition might be a workaround, but sadly, no. It fails to do
anything, as far as I can tell.