Do not set GR if we can't restore it
Since I don't know yet how to query Graphic Right (the 8-bit charset),
I've modified this program to only set Graphic Left for each of G{0,1,2,3}.

This also fixes a bug where the script was setting GR to G1 which was
an incorrect guess of mine. The VT340 actually defaults to GR => G2 at
power on.