DRCS soft-font preview, with zoom
Based on the simple fontpreview.sh, this also allows the user to pick
a character to zoom in on. Better error handling. Makes presumptions
about having a good terminfo describing the terminal, though. For
example, currently arrow keys work on xterm but not VT340 because the
the VT340 terminfo doesn't define them. (Use hjkl as a workaround for now).

Renamed to "drcspreview" since there are many types of fonts and this
only handles Dynamically Redefinable Character Sets.