Trim to bounding box width
Does not yet use offset or bounding box height.

Nor does it generate blank area if the bbx width is larger than the
actual width, but I'm not sure that's even technically allowed in a bdf.

I don't yet know if bbx height is supposed to trim/generate from the
top or bottom.

Offset appears to only matter for printing so that descenders will be
placed correctly (y-offset). I guess x-offset might be used to give a
certain character more space, maybe?