General program for timing VT terminals
Runs a program specified on the command line and shows how long it
took for it to run, including the time for the terminal to finish and
be ready again, in seconds and nanoseconds.

Uses Device Status Request 5 to figure out when the terminal is ready.
read -s -d "n" -p $'\e[5n'