Test of 80 to 1 pixel aspect ratio
A hand crafted sixel file that uses an extreme aspect ratio in the
Raster Attributes ('"') of 80:1. That is, it is 80 times taller than
it is wide. The VT340 shows this replicating each pixel in the image
vertically, causing each sixel to be 480 screen pixels high.

Note: Timing tests indicate that the VT340 takes about the same amount
of time to render no matter if the aspect ratio is 1:1 to 100000:1.