Latex with tikz graphics of projecting a mesh sphere
This is latex source which will plot Steiner circles
written by Jasper Nice, January 2025, https://tikz.net/7997-2/

It is being added to this repository as an example of how one might
view such images on a VT340 by converting it to a sixel bitmap.

%     pdflatex steiner.tex
%     convert -density 300 steiner.pdf -geometry 800x480 \
%             -background black -fill white -colorize 100% -flatten \
%             +dither -depth 4 -colors 15 sixel:- > steiner.vt340.six
%

Being mostly vector curves, it would likely be faster, or at least
more interesting, to plot using ReGIS, but there is no easy way to
convert tikz to ReGIS at the moment.